Add config file to .gitignore
One of the default config file locations is $CWD/neko_config.yml which
by default would be the repository root directory if you're running
tests locally. We don't want to accidently commit a config file as it'll
be custom to local execution of the tests. This commit adds the file to
the .gitignore to prevent this from happening.