Split out tox job for neko tests from unittests and fix issues
This commit splits out the tox job configuration for the neko
integration tests from the unittests. Previously the default tox
job was running unittests and the integration tests. This was less than
ideal because sometimes you want to just run unittests and not the neko
test suite. In doing this split some issues in the recent changes made
to the neko framework were identified and fixed as part of this.