Add component attr decorator
This commit adds a new test decorator @component_attr which enables test
authors to annotate a test method with a list of components (as strings)
on individual test methods. The expectation is that these are set on
each test method to describe which components are exercised by the test
method so that users can filter tests easily to the set of components
needed.