Add basic test guidelines to repository
In preparation for making a big push on adding new tests to the repo
it's important to have guildelines and documentation on writing tests.
This commit adds some initial documentation on this to ensure we've
got a consistent standard of expectations on what types of tests we add
to the neko test suite. This will be a living doc moving forward though
and nothing in it is set in stone. I fully expect it to grow and change
as we use qiskit neko more moving forward.