Replace ghz state quantum info test with qst experiments test
The test_ghz_circuit_quantum_info test which was added from the
qiskit-tutorials repo was not a good integration test because it was
isolated to just qiskit-terra code and didn't exercise things
wholistically. To address this that test has been removed and is
replaced by a quantum state tomography test using qiskit-experiments.
This new test runs the state tomography experiment in qiskit-experiments
to reconstruct a quantum state from measurement data after running a ghz
state circuit on a backend. The density matrix calculated is then
compared to an ideal density matrix from quantum info (similar to what
the removed test did) and the fidelity is computed. The test asserts
that the fidelity for a 3 qubit ghz state is >55%.