Move custom action definition to repo root
Using the custom action from an external repo does not work unless the
action.yml file is defined in the repository root. The actions dir is
for locally defined custom actions. This commit moves the custom action
file to the repo root accordingly so that we're able to leverage it in
other qiskit projects (which is the intent behind the custom action).