Add missing required shell arguments from custom action
The custom github action was missing a required parameter "shell" to
specify the shell to run commands in from several steps in the action.
This causes failures when trying to use the action as the action isn't
valid. This commit fixes this issue and updates the action to specify
all the commands are run in bash.