Fix lint issues with qiskit-aer >=0.11.0 (#14)
In the qiskit-aer 0.11.0 release the Python namespace of the project
moved from qiskit.providers.aer to qiskit_aer. For backwards
compatibility the old import path, qiskit.providers.aer, continues to
work and will redirect to the new name. However, pylint's static
analysis is not able to follow the metaimporter used to do this because
it happens dynamically at import time. To correct this issue, this
commit updates the import statement for aer to fix the lint job moving
forward and also migrate the neko repo to the up-to-date usage.