Add tests for primtives usage with algorithms and aer (#13)
* Add tests for primtives usage with algorithms and aer

This commit adds some new tests that exercise the primtives API with
algorithms is used. In Qiskit Terra 0.22.0 breaking API changes merged
to the primitives API, while it is too late to roll them back we can
prevent such user impacting breaking changes in the future by ensuring
we have integration test coverage in qiskit-neko. Ths commit adds tests
that use the BackendSampler and BackendEstimator primitives with
SamplingVQE and VQE to test the backend interface, and then also to
ensure we're validating the cross-project compatibility with the
abstract primitive interface tests that use the specific Aer
implementation of the primitives API is added. This will ensure that no
change to qiskit-terra can merge which will break an external
implementation of the primitives API in the future.

A follow on to this is to add support for configurable primitives
similar to what we already have with backends. This will enable adding
test jobs that use a provider's native primitive implenetations for
primitive tests.

* Update qiskit_neko/tests/primitives/test_vqe.py

* Add meaningless license header to empty file to fix lint