Set seed for primitive based machine learning tests (#16)
In the recently added primitive based machine learning tests we forgot
to set a seed on the Aer primtives. This has been causing a non-zero
failure rate in CI as the the rng values cause poorer quality results
than expected by the test tolerance. This commit sets a fixed seed via
run option in the aer primitive initialization to ensure we're running
with a stable rng output and the tests will pass consistently.