Update setuptools as part of action definition (#17)
In CI for downstream projects the neko action has been failing on a
setuptools error. This only started recently, likely from a change in
version published in the base image. This commit attempts to fix this by
updating the setuptools and pip version to the latest release as part of
the action definition.