Update tox configuration for tox 4 compatibility (#19)
This commit updates the tox configuration to support running with tox4.
Tox 4, is a major rewrite of the internals of tox and several things
behave quite differently and the configuration format had a large number
of breaking changes. This commit should hopefully enable running tox 3
or tox 4 without issue now.