Update tests for Qiskit 1.0
This commit updates the execute tests to remove the use of execute()
which was removed in Qiskit 1.0 after being deprecated in 0.46. How this
managed to merge, I'm not entirely clear on because it should have
failed CI on the removal PR, so there is likely a configuration issue in
the neko custom action as well that will need to be fixed.