Remove Backend primtive v1 dependent tests
Right now there are tests that rely on the `BackendSampler`, `Sampler`,
`Estimator` and `BackendEstimator` classes. These tests are testing the
full path execution workflow with primitives for qiskit-algorithms and
application modules with v1 implementations as input to these modules.
However, these v1 primitive implementations are deprecated and will be
removed in qiskit 2.0. This makes these test infeasible moving forward
because qiskit-algorithms only supports the primitives v1 interface, so
while there is a v2 interface available for the backend primitives in
qiskit these can't be used with algorithms anymore. So this commit
deletes the tests in questions to unblock the qiskit removal PR.