Update ci configuration to use supported python version
The repo hasn't been updated in some time as the tests were working
as-is. But in the meantime the CI configuration has bitrotted slightly.
The tests are all built around using python 3.8 which is now unsupported
both by qiskit and cPython. This commit updates the configuration to fix
this and unblock CI.