Update ci configuration to use supported python version (#51)
* Update ci configuration to use supported python version

The repo hasn't been updated in some time as the tests were working
as-is. But in the meantime the CI configuration has bitrotted slightly.
The tests are all built around using python 3.8 which is now unsupported
both by qiskit and cPython. This commit updates the configuration to fix
this and unblock CI.

* Remove setuptools pinning

* Uncap sphinx theme

* Remove unused docs requirements

* Use ruff instead of pylint for lint

* Remove one last setuptools pin from docs ci config