Handle import error in qiskit-ibm-runtime
With the pending removals in Qiskit 2.0.0 qiskit-ibm-runtime will import
error because of its use of private enums that were moved as part of the
development of Qiskit 2.0.0. The runtime provider should not be using
these attributes but in order to unblock the release of qiskit 2.0.0
this commit updates the aer plugin to tolerate the inability to import
from qiskit-ibm-runtime. This should unblock qiskit development because
qiskit doesn't use the backend selection in config to run on a noisy
backend (or I couldn't find where if it does). This should mean we only
need aer to run tests that rely on a backend.